backend/requirements.txt
aioredis==2.0.1
fastapi==0.115.3
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
python-dotenv==1.0.1
setuptools==68.2.2
uvicorn==0.30.6